fix: :bug: Avoid layer.selected_track is None
Having a None selected track_id and trying to map it through the colormap caused the error. To avoid this, we create a new track id if the selected track id is None. We also update the selected track when we update the selected node. There is still a problem with not using our custom new label function when you press m, but this is a separate issue.